15_Production_Ready_Project/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.0
faiss-cpu==1.8.0
rank-bm25==0.2.2
sentence-transformers==3.0.0
pypdf==4.2.0
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
tiktoken==0.7.0
